Use Int for port to make API more consistent
Motivation:

In ClientBootstrap we use Int to represent the port but in ServerBootstrap we use Int32. As swift usually always use Int we should just use Int everywhere to represent the port.

Modifications:

Change Int32 to Int for port

Result:

More consistent API

+++ Sources/NIO/GetaddrinfoResolver.swift
@@ -80,7 +80,7 @@ internal class GetaddrinfoResolver: Resolver {
         var info: UnsafeMutablePointer<addrinfo>?
 
         guard getaddrinfo(host, String(port), nil, &info) == 0 else {
-            self.fail(error: SocketAddressError.unknown(host: host, port: Int32(port)))
+            self.fail(error: SocketAddressError.unknown(host: host, port: port))
             return
         }
